customer-support-agent/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-chroma
chromadb
sentence-transformers
pydantic
requests
openai
groq
langchain-huggingface